+++ docs/changelog.md
@@ -5,6 +5,7 @@ Unreleased
 - A prompt allowing to restart the Action Server will appear now when an environment file is modified in an Action package with active Action Server.
 - Add UI element to open the metadata file for Action Package
 - Fix build Action Package command to add selection step to select the Action Package in case there are multiple in workspace
+- Prompt for metadata file update if action files have changed
 
 New in 2.1.0 (2024-06-25)